fix(http): fix edge case error when importing a large CSV file
Reasoning:
This is the buggy line: `int diff = ((int) (nextOffset - offset)) - bytesToRead;`
`nextOffset` and `offset` are 48 bits numbers stored inside 64 bits long.

When subtracted from each other the result can produce be more than 32 bits.
When narrow-casting the result to `int` the higher bits are truncated.

If you are especially unlucky then the truncated `diff` value is 0 or 1
and the coalescing algo will coalesce lines which should not be coalesced.

+++ core/src/main/java/io/questdb/cutlass/text/CopyTask.java
@@ -1098,8 +1098,9 @@ public class CopyTask {
                             long nextOffset = nextLengthAndOffset & MASK;
 
                             // line indexing stops on first EOL char, e.g. \r, but it could be followed by \n
-                            int diff = ((int) (nextOffset - offset)) - bytesToRead;
-                            int nextBytesToRead = diff + nextLineLength;
+                            long diff = nextOffset - offset - bytesToRead;
+                            int nextBytesToRead = ((int) diff) + nextLineLength;
+
                             if (diff > -1 && diff < 2 && addr + bytesToRead + nextBytesToRead <= lim) {
                                 bytesToRead += nextBytesToRead;
                                 additionalLines++;
@@ -1183,10 +1184,10 @@ public class CopyTask {
                         long nextOffset = nextLengthAndOffset & MASK;
 
                         // line indexing stops on first EOL char, e.g. \r, but it could be followed by \n
-                        int diff = ((int) (nextOffset - offset)) - bytesToRead;
-                        int nextBytesToRead = diff + nextLineLength;
+                        long diff = nextOffset - offset - bytesToRead;
+                        int nextBytesToRead = (int) (diff + nextLineLength);
                         if (diff > -1 && diff < 2 && bytesToRead + nextBytesToRead <= fileBufSize) {
-                            bytesToRead += diff + nextLineLength;
+                            bytesToRead += ((int) diff) + nextLineLength;
                             additionalLines++;
                         } else {
                             break;
